requirement.txt
fastapi>=0.110.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
tortoise-orm[aiomysql]>=0.21.0
PyJWT>=2.8.0
uvicorn>=0.29.0
itsdangerous>=2.1.0
jinja2>=3.1.0
python-multipart>=0.0.9